Operator spec WIP
