--- src/golang/lib/collections/operator/metric/metric.go
@@ -1,9 +1,0 @@
-package metric
-
-import (
-	"github.com/aqueducthq/aqueduct/lib/collections/operator/function"
-)
-
-type Metric struct {
-	Function function.Function `json:"function"`
-}
