--- src/golang/lib/collections/operator/system_metric/system_metric.go
@@ -1,6 +1,0 @@
-package system_metric
-
-// The value of a system metric must be JSON serializable.
-type SystemMetric struct {
-	MetricName string `json:"metric_name"`
-}
