+++ src/golang/lib/models/shared/operator/connector/extract_params.go
@@ -1,11 +1,18 @@
 package connector
 
 import (
-	gh_types "github.com/aqueducthq/aqueduct/lib/collections/operator/connector/github"
+	gh_types "github.com/aqueducthq/aqueduct/lib/models/shared/operator/connector/github"
 )
 
 type ExtractParams interface {
 	isExtractParams()
+}
+
+type MongoDBExtractParams struct {
+	Collection string `json:"collection"`
+	// `QuerySerialized` is the serialized {args, kwargs} passed to
+	// mongo's `.find()` API.
+	QuerySerialized string `json:"query_serialized"`
 }
 
 type RelationalDBExtractParams struct {
@@ -71,6 +78,8 @@ func (*RelationalDBExtractParams) isExtractParams() {}
 func (*SalesforceExtractParams) isExtractParams() {}
 
 func (*S3ExtractParams) isExtractParams() {}
+
+func (*MongoDBExtractParams) isExtractParams() {}
 
 // `CastToRelationalDBExtractParams` performs a 'casting' from params to `*RelationalDBExtractParams`.
 // This is useful for cases where we need to explicitly access relational DB information for extract.
