+++ src/golang/lib/models/shared/operator/connector/load_test.go
@@ -5,7 +5,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/aqueducthq/aqueduct/lib/collections/integration"
+	"github.com/aqueducthq/aqueduct/lib/models/shared"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
@@ -13,8 +13,8 @@ import (
 func TestMarshalAndUnmarshallLoad(t *testing.T) {
 	postgresParams := generateLoadPostgresParams()
 	originalLoad := Load{
-		Service:       integration.Postgres,
-		IntegrationID: uuid.New(),
+		Service:       shared.Postgres,
+		IntegrationId: uuid.New(),
 		Parameters:    postgresParams,
 	}
