+++ src/golang/lib/models/shared/operator/metric/metric.go
@@ -1,8 +1,6 @@
 package metric
 
-import (
-	"github.com/aqueducthq/aqueduct/lib/collections/operator/function"
-)
+import "github.com/aqueducthq/aqueduct/lib/models/shared/operator/function"
 
 type Metric struct {
 	Function function.Function `json:"function"`
