+++ src/golang/lib/models/views/operator.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/aqueducthq/aqueduct/lib/collections/operator"
-	"github.com/aqueducthq/aqueduct/lib/collections/operator/connector"
+	"github.com/aqueducthq/aqueduct/lib/models/shared/operator/connector"
 	"github.com/google/uuid"
 )
